test: move image_dir export tests into test_serialization.py
Relocate the four image_dir / image_uri_prefix export tests from a
dedicated module into tests/test_serialization.py per review.
